requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-cloud-firestore==2.11.4
numpy==1.25.2
opencv-python-headless==4.10.0.66
cmake==3.27.4
dlib==19.24.2
face_recognition==1.3.0
